// src/app/precisa/mensajeria/programar-mensaje/model/FiltroProgramacion.ts
export class FiltroProgramacion {
    constructor(){

    }

    IdProgramacion:     number;
    IdMensaje:          number;
    TipoMensaje:        string;    
    Responsable:        string;
    IdResponsable:      number;
    Documento:          string;
    TipoDocumento:      string;
    Codigo:             string;
    FechaInicio:        Date;
    FechaFin:           Date;
    Estado:             number;

}