// src/app/precisa/mensajeria/programar-mensaje/model/DtoProgramacion.ts
export class Dtoprogramacion {
    constructor(){

    }
    Documento:          string; // documento de persona
    TipoDocumento:      string;
    Responsable:        string; // nombre de persona
    DescTipoMensaje:    string;
    TipoMensaje:        string;
    Mensaje:            string;
    IdProgramacion:     number;
    IdMensaje:          number;
    Codigo:             string;
    IdResponsable:      number; // codigo de persona
    TipoProgramacion:   string;
    FechaInicio:        Date;
    FechaFin:           Date;
    HoraInicio:         Date;
    HoraFin:            Date;
    Total:              number;
    TotalEnviado:       number;
    TotalExito:         number;
    TotalError:         number;
    Estado:             number;
    UsuarioCreacion:    string;
    UsuarioModificacion: string;
    FechaCreacion:       Date;
    FechaModificacion:   string;
    IpCreacion:          string;
    IpModificacion:      string;
    DesEstado:           string;

    num:                 number;    

}